Don't detect Opera in PATH to avoid false positives
The Opera binary is just called Launcher.exe, which is very common as an exe name, so can match plenty of other programs (as one example: Intellij). Checking for this in PATH makes false positives like that extremely likely, while the other search sources are far more likely to be accurate.